Refactor JSON editor to modularize template expression utilities, enhancing code organization and reusability. Introduce common functions for validation, completion, and rendering of template pills, while improving cursor navigation and state management.
